+++ src/index.ts
@@ -1,5 +1,5 @@
 export { Service, Resource, IElement, IResourceUpdate, StatusCode } from "./rsiPlugin";
-export { Xobject } from "./xObject";
+export { XObject } from "./xobject";
 export { ElementResponse, CollectionResponse, Response } from "./response";
 export { SchemaPlugin } from "./schemaPlugin";
 export { SchemaResource} from "./schemaResource";