+++ src/xObject.ts
@@ -1,16 +1,12 @@
 import { v4 } from "uuid";
 
-export abstract class Xobject {
-  public readonly id: string;
-  public name: string;
+export abstract class XObject {
 
   /**
    * @constructor with default values
-   * @param uuid {string} use id or if left empty an v4 uuid is auto generated
+   * @param id {string} use id or if left empty an v4 uuid is auto generated
    * @param name {string} use name of if left empty an empty string is created
    */
-  constructor(uuid: string = v4(), name: string = "") {
-    this.id = uuid;
-    this.name = name;
+  constructor(public readonly id: string = v4(), public name: string = "") {
   }
 }